refactor(parsing): Move HttpParser into Parsing
