+++ Paracord.Core.UnitTests/Parsing/Http/HttpParser/DeserializeRequestTests.cs
@@ -1,6 +1,7 @@
 using System.Text;
 
 using Paracord.Core.Http;
+using Paracord.Core.Parsing.Http;
 
 namespace Paracord.Core.UnitTests.Http.HttpParserTests
 {
