+++ Paracord.Core/Listener/HttpListener.cs
@@ -2,6 +2,7 @@ using System.Net.Sockets;
 using System.Security.Cryptography.X509Certificates;
 
 using Paracord.Core.Http;
+using Paracord.Core.Parsing.Http;
 using Paracord.Shared.Models.Http;
 using Paracord.Shared.Models.Listener;
