+++ Paracord.Core/Parsing/Http/HttpParser.Deserialize.cs
@@ -1,39 +1,13 @@
 using System.Text;
 
+using Paracord.Core.Http;
 using Paracord.Shared.Models.Http;
 using HttpMethod = Paracord.Shared.Models.Http.HttpMethod;
 
-namespace Paracord.Core.Http
+namespace Paracord.Core.Parsing.Http
 {
-    public static class HttpParser
+    public static partial class HttpParser
     {
-        /// <summary>
-        /// The current version of HTTP supported by Paracord.
-        /// /// </summary>
-        public static HttpVersion CurrentHttpVersion { get; } = new HttpVersion(1, 1);
-
-        /// <summary>
-        /// Byte sequence indicating the end-of-line, according to the HTTP/1.1 specification.
-        /// </summary>
-        /// <seealso href="https://datatracker.ietf.org/doc/html/rfc9112#name-message-parsing">Reference</seealso>
-        private static readonly byte[] EndOfLine = new byte[] { 0x0D, 0x0A };
-
-        /// <summary>
-        /// Byte sequence indicating the end-of-line, according to the HTTP/1.1 specification, in string format.
-        /// </summary>
-        /// <seealso href="https://datatracker.ietf.org/doc/html/rfc9112#name-message-parsing">Reference</seealso>
-        private static readonly string EndOfLineString = Encoding.ASCII.GetString(EndOfLine);
-
-        /// <summary>
-        /// Byte sequence indicating the end of the HTTP header and start of body.
-        /// </summary>
-        private static readonly byte[] ContentDelimiter = new byte[] { 0x0D, 0x0A, 0x0D, 0x0A };
-
-        /// <summary>
-        /// Byte sequence indicating the end of the HTTP header and start of body, in string format.
-        /// </summary>
-        private static readonly string ContentDelimiterString = Encoding.ASCII.GetString(ContentDelimiter);
-
         /// <summary>
         /// Parse the specified Byte-array into an HTTP request.
         /// </summary>
@@ -168,48 +142,6 @@ namespace Paracord.Core.Http
             }
 
             request.Body.Write(bodyData);
-        }
-
-        /// <summary>
-        /// Parse the specified <see cref="HttpResponse" /> into a valid HTTP response.
-        /// </summary>
-        /// <param name="response">The response to serialize.</param>
-        /// <returns>A valid HTTP response.</returns>
-        public static byte[] SerializeResponse(HttpResponse response)
-        {
-            using MemoryStream content = new MemoryStream();
-
-            Action<string> WriteStringContent = (value) =>
-            {
-                content.Write(Encoding.ASCII.GetBytes(value));
-            };
-
-            // Append status-line
-            WriteStringContent($"{CurrentHttpVersion.ToString()} {((int) response.StatusCode)}");
-
-            // Append headers
-            foreach(var key in response.Headers.AllKeys)
-            {
-                WriteStringContent(HttpParser.EndOfLineString + $"{key}: {response.Headers[key]}");
-            }
-
-            // Append cookies
-            foreach(var key in response.Cookies.AllKeys)
-            {
-                WriteStringContent(HttpParser.EndOfLineString + $"{HttpHeaders.SetCookie}: {key}={response.Headers[key]}");
-            }
-
-            // Append content-delimiter
-            WriteStringContent(HttpParser.ContentDelimiterString);
-
-            // Append body, if present
-            if(response.Body.Length > 0)
-            {
-                response.Body.Seek(0, SeekOrigin.Begin);
-                response.Body.CopyTo(content);
-            }
-
-            return content.ToArray();
         }
     }
 }
